Fixed benchmark
git-svn-id: https://svn.origo.ethz.ch/zoodb/trunk@178 b278144e-e56a-4748-98cb-974f122dab01

+++ src/org/zoodb/jdo/internal/server/index/AbstractPagedIndex.java
@@ -291,7 +291,7 @@ public abstract class AbstractPagedIndex extends AbstractIndex {
 //		}
 		
 		
-		protected final AbstractIndexPage readOrCreatePage(short pos, boolean allowCreate) {
+		protected final AbstractIndexPage readOrCreatePage(int pos, boolean allowCreate) {
 			AbstractIndexPage page = leaves[pos];
 			if (page != null) {
 				//page is in memory
